Improve the lists

+++ components/ListComponents/List.tsx
@@ -1,4 +1,4 @@
-import { FlatList, StyleSheet, View } from 'react-native';
+import { FlatList, StyleSheet, View, ActivityIndicator } from 'react-native';
 import ListItem from './ListItem';
 import { ListItemType } from '../../constants/listItemType';
 import EmptyList from '../../baseComponents/EmptyList';
@@ -7,31 +7,40 @@ import ParagraphText from '../../baseComponents/ParagraphText';
 interface ListProps {
   data: ListItemType[];
   error: string;
+  isLoading: boolean;
   deleteItem: (item: ListItemType) => void;
 }
 
 const List = (props: ListProps) => {
-  const { data, error, deleteItem } = props;
+  const { data, error, isLoading, deleteItem } = props;
 
   const Divider = () => <View style={styles.itemSeparator} />;
 
   const renderContent = () => {
+    if (isLoading) {
+      return <ActivityIndicator size='large' color='#6c63ff' />;
+    }
+
     if (error) {
       return <ParagraphText>{error}</ParagraphText>;
-    } else {
-      return (
-        <FlatList
-          data={data}
-          showsVerticalScrollIndicator={false}
-          renderItem={({ item }) => (
-            <ListItem {...item} deleteItem={deleteItem} />
-          )}
-          ItemSeparatorComponent={() => <Divider />}
-          ListEmptyComponent={() => <EmptyList />}
-          style={{ backgroundColor: '#ffffff' }}
-        />
-      );
     }
+
+    if (data.length === 0) {
+      return <EmptyList />;
+    }
+
+    return (
+      <FlatList
+        data={data}
+        showsVerticalScrollIndicator={false}
+        renderItem={({ item }) => (
+          <ListItem {...item} deleteItem={deleteItem} />
+        )}
+        ItemSeparatorComponent={() => <Divider />}
+        // ListEmptyComponent={() => <EmptyList />}
+        style={{ backgroundColor: '#ffffff' }}
+      />
+    );
   };
 
   return <View style={{ flex: 1 }}>{renderContent()}</View>;
